backend/pyproject.toml
[tool.poetry]
name = "free-learning-bc"
version = "0.1.0"
description = ""
authors = ["Alexander Ptakhin <me@aptakhin.name>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fl_core"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
httpx = "^0.23.1"
uvicorn = "^0.20.0"
motor = {extras = ["srv"], version = "^3.1.1"}
pydantic = {extras = ["dotenv"], version = "^1.10.2"}


[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
wemake-python-styleguide = "^0.17.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
